backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.24.0
pydantic==2.11.7
pydantic-settings==2.5.2
kafka-python==2.0.2
python-json-logger==2.0.4
psutil==5.9.6
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.9
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
openai==1.97.0
mcp==1.12.0
fastmcp==2.10.6
anthropic==0.58.2
google-generativeai==0.8.5